Fix serve issue
