+++ projects/uswds-components/src/lib/step-indicator/step-indicator.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { UsaStepIndicatorComponent } from './step-indicator.component';
-import { UsaStepIndicatorHeaderComponent } from './step-indicator-header.component';
+import { UsaStepIndicatorComponent, UsaStepIndicatorHeaderComponent } from './step-indicator.component';
